backend/requirements.txt
fastapi==0.100.0
numpy==1.23.5
openai==1.2.2
pydantic==2.0.3
python-dotenv==1.0.0
PyYAML==6.0
tiktoken==0.4.0
urllib3==2.0.3
uvicorn==0.23.0
instructor==0.4.2
scikit-learn==1.2.2
pytest==7.4.0
tenacity==8.2.2
boto3==1.28.58
astor==0.8.1
rdflib==7.0.0
anthropic==0.23.1